requirements.txt
# Core dependencies
# pyo3>=0.19.0  # Only needed for building Rust extensions
# maturin>=1.0.0  # Only needed for building Rust extensions

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Logging and monitoring
psutil>=5.9.0
rich>=13.0.0

# Async and concurrency
asyncio-mqtt>=0.11.0
aiofiles>=23.0.0

# JSON and serialization
pydantic>=2.0.0
jsonschema>=4.17.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Excel support for results export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Additional statistical analysis
statsmodels>=0.14.0

# System monitoring
psutil>=5.9.0

# Progress bars
tqdm>=4.65.0

# Color output
colorama>=0.4.6
